Some fixes on types

+++ 2048/2048/Feature.cs
@@ -8,6 +8,6 @@ namespace _2048
 {
     interface Feature
     {
-        float compute(float[][] board);
+        float compute(int[][] board);
     }
 }
